requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
qdrant-client==1.7.0
transformers==4.37.2
torch==2.2.0
numpy==1.26.3
python-multipart==0.0.9
jinja2==3.1.3
apscheduler==3.10.4 
gitpython==3.1.42
sentence-transformers==4.1.0